feat(projects/products/unstable/simkernel/backend,projects/products/unstable/simkernel/ui): implement scenario query replay flows
Part of #897

+++ projects/products/unstable/simkernel/backend/src/protocol/request.rs
@@ -13,21 +13,21 @@ pub enum Request {
     },
     NewRun {
         pack_kind: String,
-        seed: u64,
-        ticks: u64,
-        turns: u64,
-        ticks_per_turn: u64,
+        seed: f64,
+        ticks: f64,
+        turns: f64,
+        ticks_per_turn: f64,
     },
     SubmitCommands {
         commands: Vec<String>,
     },
     Step {
-        n_ticks: u64,
+        n_ticks: f64,
     },
     RunToEnd,
     GetSnapshot {
-        at_tick: u64,
-        at_turn: u64,
+        at_tick: f64,
+        at_turn: f64,
     },
     Query {
         query: String,
